nu ska allt funka

+++ ClassLibrary1/Faktura.cs
@@ -8,13 +8,13 @@ namespace BusinessLayer
     {
         //klassen faktura. StartDate  och EndDate håller koll på när fakturan startar och slutar
         public int FakturaId { get; set; }
-        public int Pris { get; set; }
+        public double Pris { get; set; }
         public  Bokning bokning { get; set; }
         public List<Bok> böcker { get; set; }
         public DateTime startDate { get; set; }
         public DateTime endDate { get; set; }
 
-        public Faktura(int fakturaId, int pris, Bokning bokning, List<Bok> böcker, DateTime startDate, DateTime endDate)
+        public Faktura(int fakturaId, double pris, Bokning bokning, List<Bok> böcker, DateTime startDate, DateTime endDate)
         {
             this.FakturaId = fakturaId;
             this.Pris = pris;
